PUBDEV-497:  parse gzip/zip files.  Added code to support parsing a single zip file or a zipped directory.
PUBDEV-1843: grid testing. Updated code to reflect changes in Python APIs and make sure all tests pass.

PUBDEV-3359: parse 1000s file.  Added pyunit test to parse 2000 files

+++ h2o-py/dynamic_tests/testdir_algos/rf/pyunit_rf_gridsearch_sorting_metrics.py
@@ -41,7 +41,7 @@ class Test_rf_gridsearch_sorting_metrics:
     diff = 1e-10            # comparison threshold
 
     curr_time = str(round(time.time()))     # store current timestamp, used as part of filenames.
-    seed = round(time.time())
+    seed = int(round(time.time()))
 
     # parameters denoting filenames of interested that store training/validation/test data sets in csv format
     training1_filename = "smalldata/gridsearch/multinomial_training1_set.csv"
